backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
email-validator==2.2.0
sqlalchemy==2.0.45
alembic==1.14.1
razorpay==2.0.0
python-dotenv==1.0.1
